testing 101

+++ tax_app/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class TaxationScheme(models.Model):
     year = models.PositiveIntegerField(unique=True)
-    tax_rate = models.DecimalField(max_digits=5, decimal_places=2)
+    tax_rate = models.PositiveIntegerField()
 
     def __str__(self):
         return f"TaxationScheme - {self.year}"
